+++ debug/example/activity/activity.js
@@ -9,11 +9,13 @@ const onReady = () => {
   const urlManager = new UrlManager(document.location.toString());
   new ElementButton('#go_right').addEventListener('click', function () {
     history.pushState('#slideR2L', 'right', `?a=b`);
-    const activityId = Activity.onStart('fvueSlideFromRight top-[200px]', 'fvueSlideToLeft');
+    const activityId = Activity.onStart('fvueSlideFromRight top-[200px] h-[400px]', 'fvueSlideToLeft left-[50%]');
+    Activity.setStateHistory(activityId);
   });
   new ElementButton('#go_bottom').addEventListener('click', function () {
     history.pushState('#slideB2T', 'bottom', `?b=c`);
     const activityId = Activity.onStart('fvueSlideFromBottom top-[140%]', 'fvueSlideToTop');
+    Activity.setStateHistory(activityId);
   });
 };
 // document ready