+++ debug/plugins/prism-es6/components/prism-elixir.js
@@ -878,6 +878,9 @@ Prism.languages.elixir = {
 
 
 
+
+
+
       // See interpolation below
     } }, { pattern: /("""|''')[\s\S]*?\1/, greedy: true, inside: {